backend/requirements.txt
flask>=3.1.2
flask-cors>=6.0.1
langchain>=0.1.0
langchain-groq>=0.0.3
langchain-community>=0.0.13
python-dotenv>=1.0.0
reportlab>=4.0.8
wordcloud>=1.9.3
matplotlib>=3.8.2
numpy>=1.26.2
wikipedia-api>=0.6.0
langgraph>=0.1.0